requirements.txt
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.6
pycurl==7.45.3
pyparsing==3.1.2
python-nmap==0.7.1
requests==2.31.0
setuptools==69.1.1
six==1.16.0
soupsieve==2.5
urllib3==2.2.1
wfuzz==3.1.0